Create time range util file
